from monitor move Body, data names into agent, Clock data names into env
