+++ test/fixture/common/util/test_dict.json
@@ -1,1 +1,5 @@
-{"a": 1, "b": 2, "c": 3}
+{
+  "a": 1,
+  "b": 2,
+  "c": 3
+}
